+++ tests/app/integrations/comp_and_pen/test_scheduled_message_helpers.py
@@ -1,5 +1,4 @@
 from decimal import Decimal
-from unittest.mock import MagicMock
 
 import pytest
